+++ docker-compose.yaml
@@ -22,7 +22,7 @@ services:
     container_name: app-init
     env_file:
       - .env
-    command: pnpx tsx src/init.ts
+    command: npx tsx src/init.ts
   app:
     image: frivsspr
     container_name: app
@@ -30,6 +30,7 @@ services:
       - .env
     depends_on:
       - app-migrate
+      - app-init
     ports:
       - 3000:3000
     restart: on-failure
