+++ src/lib/types.ts
@@ -1,28 +1,6 @@
-export interface SubjectInfo {
-  name: string;
-  code: string | null;
-  points?: number;
-  type?: SubjectType;
-  notTaught?: boolean;
-}
-
-export enum SubjectType {
-  Mandatory = 'Obvezni predmet',
-  Disciplinary = 'Strokovni izbirni predmet',
-  Directionary = 'Smerni izbirni predmet',
-  Common = 'Splošni izbirni predmet',
-}
-
-export interface Subject {
-  id: number;
-  info: SubjectInfo;
-  prerequisites?: SubjectInfo[];
-  related?: SubjectInfo[];
-  comments: Comment[];
-  ratings: number[];
-  voteCount: number;
-  semester: number;
-}
+export type { Subject, SubjectInfo, Comment } from '../base_types';
+export { SubjectType } from '../base_types';
+import type { SubjectType } from '../base_types';
 
 export interface CalculatorSubject {
   name: string;
@@ -31,11 +9,4 @@ export interface CalculatorSubject {
   type: SubjectType;
   isSelected: boolean;
   notTaught?: boolean;
-}
-
-interface Comment {
-  id: number;
-  content: string;
-  approved: boolean;
-  createdAt: string;
 }
